[fix]: add performance improvement to dashboard service

+++ backend/observatorio/src/main/java/com/upe/observatorio/projeto/dominio/envelopes/DashboardResumoVO.java
@@ -6,9 +6,9 @@ import lombok.Data;
 
 @Data
 public class DashboardResumoVO {
-	private Integer totalCourses;
+	private Long totalCourses;
 	
-	private Integer totalCampuses;
+	private Long totalCampuses;
 	
-	private Integer totalProjects;
+	private Long totalProjects;
 }
